+++ JPEG-Encoder-SIMD/lodepng.cpp
@@ -772,7 +772,7 @@ unsigned lodepng_huffman_code_lengths(unsigned* lengths, const unsigned* frequen
 	BPMNode* leaves; /*the symbols, only those with > 0 frequency*/
 
 	if (numcodes == 0) return 80; /*error: a tree of 0 symbols is not supposed to be made*/
-	if ((1u << maxbitlen) < numcodes) return 80; /*error: represent all symbols*/
+	if ((size_t(1u) << (size_t)maxbitlen) < numcodes) return 80; /*error: represent all symbols*/
 
 	leaves = (BPMNode*)lodepng_malloc(numcodes * sizeof(*leaves));
 	if (!leaves) return 83; /*alloc fail*/
@@ -1438,11 +1438,11 @@ static void updateHashChain(Hash* hash, size_t wpos, unsigned hashval, unsigned 
 {
 	hash->val[wpos] = (int)hashval;
 	if (hash->head[hashval] != -1) hash->chain[wpos] = hash->head[hashval];
-	hash->head[hashval] = wpos;
+	hash->head[hashval] = (int)wpos;
 
 	hash->zeros[wpos] = numzeros;
 	if (hash->headz[numzeros] != -1) hash->chainz[wpos] = hash->headz[numzeros];
-	hash->headz[numzeros] = wpos;
+	hash->headz[numzeros] = (int)wpos;
 }
 
 /*
@@ -1514,7 +1514,7 @@ static unsigned encodeLZ77(uivector* out, Hash* hash,
 		for (;;)
 		{
 			if (chainlength++ >= maxchainlength) break;
-			current_offset = hashpos <= wpos ? wpos - hashpos : wpos - hashpos + windowsize;
+			current_offset = (unsigned int)(hashpos <= wpos ? wpos - hashpos : wpos - hashpos + windowsize);
 
 			if (current_offset < prev_offset) break; /*stop when went completely around the circular buffer*/
 			prev_offset = current_offset;
@@ -3445,7 +3445,7 @@ unsigned lodepng_convert(unsigned char* out, const unsigned char* in,
 		for (i = 0; i != palsize; ++i)
 		{
 			unsigned char* p = &mode_out->palette[i * 4];
-			color_tree_add(&tree, p[0], p[1], p[2], p[3], i);
+			color_tree_add(&tree, p[0], p[1], p[2], p[3], (unsigned int)i);
 		}
 	}
 
@@ -4248,7 +4248,7 @@ static unsigned readChunk_tEXt(LodePNGInfo* info, const unsigned char* data, siz
 
 		string2_begin = length + 1; /*skip keyword null terminator*/
 
-		length = chunkLength < string2_begin ? 0 : chunkLength - string2_begin;
+		length = (unsigned int)(chunkLength < string2_begin ? 0 : chunkLength - string2_begin);
 		str = (char*)lodepng_malloc(length + 1);
 		if (!str) CERROR_BREAK(error, 83); /*alloc fail*/
 
@@ -4296,7 +4296,7 @@ static unsigned readChunk_zTXt(LodePNGInfo* info, const LodePNGDecompressSetting
 		string2_begin = length + 2;
 		if (string2_begin > chunkLength) CERROR_BREAK(error, 75); /*no null termination, corrupt?*/
 
-		length = chunkLength - string2_begin;
+		length = (unsigned int)(chunkLength - string2_begin);
 		/*will fail if zlib error, e.g. if length is too small*/
 		error = zlib_decompress(&decoded.data, &decoded.size,
 			(unsigned char*)(&data[string2_begin]),
@@ -4376,7 +4376,7 @@ static unsigned readChunk_iTXt(LodePNGInfo* info, const LodePNGDecompressSetting
 		/*read the actual text*/
 		begin += length + 1;
 
-		length = chunkLength < begin ? 0 : chunkLength - begin;
+		length = (unsigned int)(chunkLength < begin ? 0 : chunkLength - begin);
 
 		if (compressed)
 		{
@@ -5316,7 +5316,7 @@ static unsigned filter(unsigned char* out, const unsigned char* in, unsigned w, 
 		{
 			for (type = 0; type != 5; ++type)
 			{
-				unsigned testsize = attempt[type].size;
+				unsigned testsize = (unsigned int)attempt[type].size;
 				/*if(testsize > 8) testsize /= 8;*/ /*it already works good enough by testing a part of the row*/
 
 				filterScanline(attempt[type].data, &in[y * linebytes], prevline, linebytes, bytewidth, type);
